+++ mdadash/backend/analyses/__init__.py
@@ -2,11 +2,13 @@
 Module that has all the analyses widgets
 """
 
-from . import com_distance, dssp, energies, rog
+from . import com_distance, dssp, energies, janin, ramachandran, rog
 
 __all__ = [
     "energies",
     "com_distance",
     "rog",
     "dssp",
+    "ramachandran",
+    "janin",
 ]